[sw,tests] Update example chip-level test template.
This updates the example chip-level test template that can be used by
test developers as a starting point for developing device-side
chip-level test code.
